Add "storage", "skia", fix "sql" installation, OECMAKE_CXX
We now build Storage and Skia.

Skia requires "libjpeg-turbo" and will not compile with "jpeg",
but as it conflicts with "jpeg" if it has been used for other
recipes ("gdk-pixbuf", "gtk+"...), we force it at a layer level
with PREFERRED_PROVIDER (this may require rebuilding the image).

OECMAKE_CXX* variables may be overridden if we use EXTRA_OECMAKE
in the same recipe; this only happens for Skia currently, but
fix it everywhere by replacing them with CXXFLAGS/LDFLAGS so
that it does not bite us in the future.

Some headers were missing during SQL installation.

Signed-off-by: Manuel Bachmann <mbc@iot.bzh>