Add "media-blink", "content", "ui_views", "content_shell", "minizip"
We now build all required components.

We now .bbappend to zlib in order to provide "minizip", which is
required by Content. As this triggers a full rebuild, we may want to
have a standalone recipe instead...

We now provide various resources files (.js, .json, .mojom) from
various packages, which are required by Content.

Signed-off-by: Manuel Bachmann <mbc@iot.bzh>