Fix Mojo cross-compilation, out-of-tree patches rebuild
The "python-jinja2" recipe we provide, used by all
Mojo-dependent recipes to generate bindings at configure
time, is not really working. So let us remove it and
have Jinja2 directly in Mojo as "third_party/jinja2"
(upstream Chromium does this).

The way we handle patches for code above "${S}" was
prone to fail each time we re-configured the same package,
so let us have a workaround.

Signed-off-by: Manuel Bachmann <mbc@iot.bzh>